Move ChainRulesCore support to extension
