+++ src/density-core.jl
@@ -35,15 +35,6 @@ end
 
 _checksupport(cond, result) = ifelse(cond == true, result, oftype(result, -Inf))
 
-import ChainRulesCore
-@inline function ChainRulesCore.rrule(::typeof(_checksupport), cond, result)
-    y = _checksupport(cond, result)
-    function _checksupport_pullback(ȳ)
-        return NoTangent(), ZeroTangent(), one(ȳ)
-    end
-    y, _checksupport_pullback
-end
-
 export unsafe_logdensityof
 
 # https://discourse.julialang.org/t/counting-iterations-to-a-type-fixpoint/75876/10?u=cscherrer
