+++ src/transport.jl
@@ -135,9 +135,6 @@ end
     return static(10)
 end
 
-_origin_depth_pullback(ΔΩ) = NoTangent(), NoTangent()
-ChainRulesCore.rrule(::typeof(_origin_depth), ν) = _origin_depth(ν), _origin_depth_pullback
-
 # If both both measures have no origin:
 function _transport_between_origins(ν, ::StaticInteger{0}, ::StaticInteger{0}, μ, x)
     _transport_with_intermediate(ν, _transport_intermediate(ν, μ), μ, x)
